Fix issue with stale clientConn
The tranport does not check if the IP address of a node
has changed and returns the connection created with outdated info.
This happens in environments with dynamic IPs such a docker
and Kubernetes.
This fix adds an additional check to ensure the the target IP is up
to date.